implementacion de lobby
