+++ src/components/room/lobby/TabInfo.vue
@@ -1,0 +1,1 @@
+<template></template>
