+++ src/components/room/lobby/TabUsers.vue
@@ -1,0 +1,1 @@
+<template></template>
