+++ src/pages/index.astro
@@ -1,10 +1,9 @@
 ---
 import Layout from "../layouts/Layout.astro";
-import WebSocketDemo from "../components/test/WebSocketDemo.vue";
 ---
 
 
 <Layout title="Suffgo">
-  <WebSocketDemo client:load>
+  <h1>bienvenido</h1>
 </Layout>
